+++ src/backend/Application/Features/Posts/Commands/DeletePost/DeletePostCommandHandler.cs
@@ -9,9 +9,11 @@ namespace Application.Features.Posts.Commands.DeletePost
     public sealed class DeletePostCommandHandler : IRequestHandler<DeletePostCommand, Result<bool>>
     {
         private readonly IUnitOfWork _unitOfWork;
-        public DeletePostCommandHandler(IUnitOfWork unitOfWork)
+        private readonly IMedia _media;
+        public DeletePostCommandHandler(IUnitOfWork unitOfWork, IMedia media)
         {
             _unitOfWork = unitOfWork;
+            _media = media;
         }
         public async Task<Result<bool>> Handle(DeletePostCommand request, CancellationToken cancellationToken)
         {
@@ -20,6 +22,11 @@ namespace Application.Features.Posts.Commands.DeletePost
             if (post == null)
             {
                 return Result<bool>.ResultFailures(ErrorConstants.NotFoundWithId(request.Id));
+            }
+            var DeleteImageResult = await _media.DeleteImageAsync(post.PublicIdImage);
+            if (DeleteImageResult.IsSuccess is false)
+            {
+                return Result<bool>.ResultFailures(DeleteImageResult.Errors);
             }
             repo.Delete(post);
             await _unitOfWork.Commit();
